// src/theme/types.ts
export type Gradients = {
  bubblegum: string;
};

export type Colors = {
  primary: string;
  primaryBright: string;
  primaryDark: string;
  secondary: string;
  tertiary: string;
  success: string;
  failure: string;
  warning: string;
  contrast: string;
  invertedContrast: string;
  input: string;
  inputSecondary: string;
  background: string;
  backgroundDisabled: string;
  text: string;
  textDisabled: string;
  textSubtle: string;
  textLight: string;
  textDark: string;
  borderColor: string;
  card: string;

  // Gradients
  gradients: Gradients;
 
  // Brand colors
  binance: string;
};